+++ client/python/vstimd/system/__init__.py
@@ -1,6 +1,8 @@
 from .system_client import SystemClient
 from .system_models import (
     Camera3D,
+    InputAxisInfo,
+    InputDeviceInfo,
     Lighting3D,
     CapturedFrame,
     DeferredModeStatus,
@@ -13,6 +15,8 @@ __all__ = [
     "SystemClient",
     "CapturedFrame",
     "Camera3D",
+    "InputAxisInfo",
+    "InputDeviceInfo",
     "Lighting3D",
     "DeferredModeStatus",
     "ServerInfo",