+++ server/src/ipc/dispatch.rs
@@ -229,6 +229,7 @@ fn command_summary(req: &proto::Request) -> String {
         Some(request::Body::QueryCamera(_)) => "QueryCamera".into(),
         Some(request::Body::SetLighting(_)) => "SetLighting".into(),
         Some(request::Body::QueryLighting(_)) => "QueryLighting".into(),
+        Some(request::Body::ListInputDevices(_)) => "ListInputDevices".into(),
         Some(request::Body::SetNavSpeed(c)) => {
             format!("SetNavSpeed({}, {:.1}cm/s)", c.handle, c.speed_cm_per_s)
         }
@@ -329,6 +330,7 @@ impl SceneState {
             request::Body::QueryCamera(_) => self.cmd_query_camera(),
             request::Body::SetLighting(cmd) => self.cmd_set_lighting(cmd),
             request::Body::QueryLighting(_) => self.cmd_query_lighting(),
+            request::Body::ListInputDevices(_) => self.cmd_list_input_devices(),
             request::Body::CreatePolygon(_) => err(
                 proto::ErrorCode::NotSupported,
                 "CreatePolygon is not yet implemented",
@@ -414,6 +416,7 @@ impl SceneState {
             | request::Body::QueryCamera(_)
             | request::Body::SetLighting(_)
             | request::Body::QueryLighting(_)
+            | request::Body::ListInputDevices(_)
             | request::Body::CreatePolygon(_)
             | request::Body::SetBackground(_)
             | request::Body::SetDeferredMode(_)